+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `pyrgg.engines.erdos_reyni` module
 ### Changed
+- `logger` function format for `erdos_reyni_gilbert` changed
+- `log` function added
 - `README.md` modified
 - `build_exe.bat` modified
 - `Python 3.13` added to `test.yml`